+++ test/math.js
@@ -9,12 +9,6 @@ describe('Math', function() {
 
 	var originalNumbers = [17, 17, 3, 2, 4, 8, 1, 0, 3, 1, 7, 12, 6, 13, 2, 14, 6]; // Sum is 116
 
-	describe('.toSource()', function() {
-		it('should return the source code representation of the Math object', function() {
-			assert.equal(Math.toSource(), 'Math');
-		});
-	});
-
 	describe('.median(numbers)', function() {
 
 		it('should return the median value of the numbers without changing the original array', function() {